create EndToEndSQLOptimizerSuite
